Adding a "use_ssl" switch
There is the use case, where some want to use the api on the same
host, where the application runs, while using localhost as hostname.
It is okay there to not speak ssl in some cases, which makes the thing
with "localhost" much easier, so I added a use_ssl-switch which is true
on default. You can pass it in when initializing the ZabbixAPI-Object.